Refactor export/import utilities into opal.core.trade
