+++ opal/views.py
@@ -18,7 +18,7 @@ from django.utils.text import slugify
 from django.views.generic import FormView, TemplateView, View
 
 from opal import models
-from opal.core import application, detail, episodes
+from opal.core import application, detail, episodes, trade
 from opal.core.patient_lists import PatientList, TabbedPatientListGroup
 from opal.core.subrecords import (
     episode_subrecords, get_subrecord_from_api_name,
@@ -306,23 +306,8 @@ class RawTemplateView(LoginRequiredMixin, TemplateView):
         return super(RawTemplateView, self).get(*args, **kw)
 
 
-def remove_key(d, key):
-    """
-    Remove the given key from the given dictionary recursively
-    """
-    for k, v in d.iteritems():
-        if k == key:
-            continue
-
-        if isinstance(v, collections.Mapping):
-            yield k, dict(remove_key(v, key))
-        elif isinstance(v, list):
-            yield k, [dict(remove_key(x, key)) for x in v]
-        else:
-            yield k, v
-
-
 class ExportEpisodeView(LoginRequiredMixin, View):
+
     def get(self, request, *args, **kwargs):
         episode_id = self.kwargs['episode_id']
 
@@ -336,7 +321,7 @@ class ExportEpisodeView(LoginRequiredMixin, View):
         data = episode.to_dict(request.user)
 
         # Remove all "id" keys from the data
-        data = dict(remove_key(data, 'id'))
+        data = dict(trade.remove_key(data, 'id'))
 
         response = json_response(data)
 
@@ -353,19 +338,11 @@ class ExportPatientView(LoginRequiredMixin, View):
         patient_id = self.kwargs['patient_id']
 
         try:
-            patient = models.Patient.objects.get(pk=patient_id)
+            data, patient = trade.patient_id_to_json(patient_id, user=request.user)
         except models.Patient.DoesNotExist:
             msg = 'Cannot find Patient with ID: {}'.format(patient_id)
             messages.error(request, msg)
             return redirect(reverse('admin:opal_patient_changelist'))
-
-        data = patient.to_dict(request.user)
-
-        # Active episode is not a concept which makes sense for a second system
-        del data['active_episode_id']
-
-        # Remove all "id" keys from the data
-        data = dict(remove_key(data, 'id'))
 
         response = json_response(data)
 
@@ -375,47 +352,6 @@ class ExportPatientView(LoginRequiredMixin, View):
         header_value = 'attachment; filename={}'.format(filename)
         response['Content-Disposition'] = header_value
         return response
-
-
-def match_or_create_patient(demographic, user):
-    """
-    Get a Patient record using demographics data
-
-    Attempt the lookup using three methods:
-        1. NHS Number
-        2. DoB, First name, & Surname
-        3. Create a new Demographic record
-    """
-    Demographics = get_subrecord_from_model_name('Demographics')
-    nhs_number = demographic.get('nhs_number')
-    if nhs_number:
-        try:
-            return Demographics.objects.get(nhs_number=nhs_number).patient
-        except Demographics.DoesNotExist:
-            pass
-
-    dob = demographic.get('date_of_birth')
-    first_name = demographic.get('first_name')
-    surname = demographic.get('surname')
-    if all([dob, first_name, surname]):
-        date_of_birth = datetime.strptime(dob, "%d/%m/%Y").date()
-        try:
-            return Demographics.objects.get(
-                date_of_birth=date_of_birth,
-                first_name=first_name,
-                surname=surname,
-            ).patient
-        except Demographics.DoesNotExist:
-            pass
-
-    # Remove data we don't want to save
-    if 'id' in demographic:
-        del demographic['id']
-
-    patient = models.Patient.objects.create()
-    d = Demographics(patient=patient)
-    d.update_from_dict(demographic, user)
-    return patient
 
 
 class ImportEpisodeView(LoginRequiredMixin, FormView):
@@ -491,19 +427,6 @@ class ImportEpisodeView(LoginRequiredMixin, FormView):
         return context
 
 
-def import_patient(data):
-    """
-    Given a datastructure representing a Patient, import that patient.
-    """
-    patient = match_or_create_patient(
-        data['demographics'][0],
-        self.request.user,
-    )
-    episodes = data.pop('episodes')
-
-    patient.update_from_dict(data, self.request.use)
-    messages.success(self.request, 'Imported {}'.format(patient))
-    return
 
 class ImportPatientView(LoginRequiredMixin, FormView):
     form_class = ImportDataForm
@@ -513,7 +436,8 @@ class ImportPatientView(LoginRequiredMixin, FormView):
     def form_valid(self, form):
         raw_data = self.request.FILES['data_file'].read()
         data = json.loads(raw_data)
-        import_patient(data)
+        trade.import_patient(data, user=self.request.user)
+        messages.success(self.request, 'Imported {}'.format(patient))
         return super(ImportPatientView, self).form_valid(form)
 
     def get_context_data(self, **kwargs):
